Quantica.jl

A Julia package for building generic tight-binding models and computing various spectral and transport properties.

Important information

This package supersedes Elsa.jl, which will soon be deprecated.

Manual

Tutorial

Examples

Reference

Current functionality

Build arbitrary lattices (periodic or bounded in any dimension and with any unit cell)

Define generic model Hamiltonians by applying a model onto a lattice

Use models with arbitrary orbital structure, spatial dependence and coordination (e.g. normal/superconducting, spin-orbit coupling, etc.)

Define parametric Hamiltonians that efficiently implement external parameters dependencies

Efficiently compute the Bloch Hamiltonian matrix at arbitrary wave vector

Compute the band structure or spectrum of a Hamiltonian.

Use advanced meshing and co-diagonalization techniques to resolve degeneracies and extract subbands

Use Order-N Kernel polynomial methods to compute spectral and transport properties efficiently

Planned functionality

Interpolated Green's functions

Landauer and Kubo formalisms

Self-consistent mean field calculations

Quantica.BoxIterator

Base.factorial

Quantica.averageKPM

Quantica.bands

Quantica.bandstructure

Quantica.bloch

Quantica.bloch!

Quantica.bravais

Quantica.combine

Quantica.combine

Quantica.densityKPM

Quantica.dims

Quantica.dosKPM

Quantica.energies

Quantica.flatten

Quantica.hamiltonian

Quantica.hopping

Quantica.hopping!

Quantica.hoppingselector

Quantica.lattice

Quantica.marchingmesh

Quantica.momentaKPM

Quantica.offdiagonal

Quantica.onsite

Quantica.onsite!

Quantica.onsiteselector

Quantica.optimize!

Quantica.parametric

Quantica.permutations

Quantica.permutations

Quantica.similarmatrix

Quantica.sites

Quantica.spectrum

Quantica.states

Quantica.states

Quantica.sublat

Quantica.supercell

Quantica.transform!

Quantica.transform!

Quantica.unitcell

Quantica.vertices

Quantica.wrap

Quantica.averageKPM — Method

averageKPM(h::AbstractMatrix, A; kBT = 0, Ef = 0, kw...)

Compute, using the Kernel Polynomial Method (KPM), the thermal expectation value <A> = Σ_k f(E_k) <k|A|k> = ∫dE f(E) Tr [A δ(E-H)] = Tr [A f(H)] for a given hermitian operator A and a hamiltonian h (see momentaKPM and its options kw for further details). f(E) is the Fermi-Dirac distribution function, kBT is the temperature in energy units and Ef the Fermi energy.

averageKPM(μ::MomentaKPM, A; kBT = 0, Ef = 0)

Same as above with the KPM momenta as input (see momentaKPM).

averageKPM(h::Hamiltonian, A::Hamiltonian; kw...)

Equivalent to averageKPM(bloch(h), bloch(A); kw...) for finite Hamiltonians (zero dimensional).

Quantica.bands — Method

bands(bs::Bandstructure)

Return a vector of all the Bands in bs.

Quantica.bandstructure — Method

bandstructure(h::Hamiltonian, mesh::Mesh; minprojection = 0.5, method = defaultmethod(h))

Compute the bandstructure of Bloch Hamiltonian bloch(h, ϕs...) with ϕs evaluated on the vertices of mesh. It is assumed that h is hermitian.

The option minprojection determines the minimum projection between eigenstates to connect them into a common subband. The option method is chosen automatically if unspecified, and can be one of the following

method                    diagonalization function
--------------------------------------------------------------
LinearAlgebraPackage()     LinearAlgebra.eigen!
ArpackPackage()            Arpack.eigs (must be `using Arpack`)

Options passed to the method will be forwarded to the diagonalization function. For example, method = ArpackPackage(nev = 8, sigma = 1im) will use Arpack.eigs(matrix; nev = 8, sigma = 1im) to compute the bandstructure.

bandstructure(h::Hamiltonian; resolution = 13, shift = missing, kw...)

Same as above with a uniform mesh = marchingmesh(h; npoints = resolution, shift = shift) of marching tetrahedra (generalized to the lattice dimensions of the Hamiltonian). Note that resolution denotes the number of points along each Bloch axis, including endpoints (can be a tuple for axis-dependent points).

Example

julia> h = LatticePresets.honeycomb() |> unitcell(3) |> hamiltonian(hopping(-1, range = 1/√3));

julia> bandstructure(h; resolution = 25, method = LinearAlgebraPackage())
Bandstructure: bands for a 2D hamiltonian
  Bands        : 8
  Element type : scalar (Complex{Float64})
  Mesh{2}: mesh of a 2-dimensional manifold
    Vertices   : 625
    Edges      : 3552

See also

marchingmesh

Quantica.bloch! — Method

bloch!(matrix, h::Hamiltonian, ...)

In-place version of bloch. Overwrite matrix with the Bloch Hamiltonian matrix of h for the specified Bloch phases ϕs (see bloch for definition and API). A conventient way to obtain a matrix is to use similarmatrix(h), which will return an AbstractMatrix of the same type as the Hamiltonian's. Note, however, that matrix need not be of the same type (e.g. it can be dense, while h is sparse).

Examples

julia> h = LatticePresets.honeycomb() |> hamiltonian(onsite(1), hopping(2));

julia> bloch!(similarmatrix(h), h, (.2,.3), 2)  # velocity operator along `bravais(h)[2]`
2×2 SparseArrays.SparseMatrixCSC{Complex{Float64},Int64} with 4 stored entries:
  [1, 1]  =  1.99001-0.199667im
  [2, 1]  =  1.96013-0.397339im
  [1, 2]  =  1.96013+0.397339im
  [2, 2]  =  1.99001-0.199667im

See also:

bloch, optimize!, similarmatrix

Quantica.bloch — Method

bloch(h::Hamiltonian{<:Lattice}, ϕs::Real...)

Build the Bloch Hamiltonian matrix of h, for the specified Bloch phases ϕs. In terms of Bloch wavevector k, ϕs = k * bravais(h), it is defined as Overwrite matrix with the Bloch Hamiltonian matrix of h, for the specified Bloch phases ϕs, defined as H(ϕs) = ∑exp(-im * ϕs' * dn) h_dn where h_dn are Bloch harmonics connecting unit cells at a distance dR = bravais(h) * dn.

bloch(h::Hamiltonian{<:Lattice})

Build the intra-cell Hamiltonian matrix of h, without adding any Bloch harmonics.

bloch(h::Hamiltonian{<:Lattice}, ϕs::NTuple{L,Real}[, axis::Int = 0])

A nonzero axis produces the derivative of the Bloch matrix respect to ϕs[axis] (i.e. the velocity operator along this axis), ∂H(ϕs) = ∑ -im * dn[axis] * exp(-im * ϕs' * dn) h_dn

bloch(matrix, h::Hamiltonian{<:Lattice}, ϕs::NTuple{L,Real}, dnfunc::Function)

Generalization that applies a prefactor dnfunc(dn) * exp(im * ϕs' * dn) to the dn harmonic.

h |> bloch(ϕs...)
h(ϕs...)

Functional forms of bloch, equivalent to bloch(h, ϕs...)

bloch(h::Hamiltonian{<:Superlattice}, ϕs::Number...)
bloch(h::Hamiltonian{<:Superlattice}, ϕs::Tuple[, axis = 0])

Build a SupercellBloch object that lazily implements the Bloch Hamiltonian in the Superlattice without actually building the matrix (e.g. for matrix-free diagonalization).

Notes

bloch allocates a new matrix on each call. For a non-allocating version of bloch, see bloch!.

Examples

julia> h = LatticePresets.honeycomb() |> hamiltonian(onsite(1), hopping(2)) |> bloch(.2,.3)
2×2 SparseArrays.SparseMatrixCSC{Complex{Float64},Int64} with 4 stored entries:
  [1, 1]  =  1.99001-0.199667im
  [2, 1]  =  1.96013-0.397339im
  [1, 2]  =  1.96013+0.397339im
  [2, 2]  =  1.99001-0.199667im

See also:

bloch!, optimize!, similarmatrix

Quantica.bravais — Method

bravais(vecs...; semibounded = false)
bravais(matrix; semibounded = false)

Create a Bravais{E,L} that adds L Bravais vectors vecs in E dimensional space, alternatively given as the columns of matrix mat. For higher instantiation efficiency enter vecs as Tuples or SVectors and mat as SMatrix.

To create semibounded lattices along some or all Bravais vectors, use semibounded. A semibounded = true makes all axes semibounded. A semibounded = (axes::Int...) indicates the indices of axes to be made semibounded. A tuple of Booleans, one per Bravais axes, semibounded = issemi::NTuple{L,Bool} is also allowed. Note that semibounded lattices always extend toward positive multiples of Bravais vectors. To invert the direction, invert the vectors.

We can scale a b::Bravais simply by multiplying it with a factor a, like a * b.

bravais(lat::Lattice)
bravais(h::Hamiltonian)

Obtain the Bravais matrix of lattice lat or Hamiltonian h

Examples

julia> bravais((1.0, 2), (3, 4))
Bravais{2,2,Float64} : set of 2 Bravais vectors in 2D space.
  Vectors : ((1.0, 2.0), (3.0, 4.0))
  Matrix  : [1.0 3.0; 2.0 4.0]

See also:

semibounded

Quantica.combine — Method

combine(hams::Hamiltonian...; coupling = missing)

Build a new Hamiltonian h that combines all hams as diagonal blocks, and applies coupling::Model, if provided, to build the off-diagonal couplings. Note that the diagonal blocks are not modified by the coupling model.

Quantica.combine — Method

combine(lats::Lattice...)

If all lats have compatible Bravais vectors, combine them into a single lattice. Sublattice names are renamed to be unique if necessary.

Quantica.densityKPM — Method

densityKPM(h::AbstractMatrix, A; resolution = 2, kw...)

Compute, using the Kernel Polynomial Method (KPM), the local spectral density of an operator A ρ_A(ϵ) = ⟨ket|A δ(ϵ-h)|ket⟩/⟨ket|ket⟩ for a given ket::AbstractVector and hamiltonian h, or the global spectral density ρ_A(ϵ) = Tr[A δ(ϵ-h)] if ket is missing. If ket is an AbstractMatrix it evaluates the trace over the set of kets in ket (see momentaKPM and its options kw for further details). A tuple of energy points xk and ρ_A values is returned where the number of energy points xk is order * resolution, rounded to the closest integer.

densityKPM(momenta::MomentaKPM; resolution = 2)

Same as above with the KPM momenta as input (see momentaKPM).

densityKPM(h::Hamiltonian, A::Hamiltonian; kw...)

Equivalent to densityKPM(bloch(h), bloch(A); kw...) for finite Hamiltonians (zero dimensional).

Quantica.dims — Method

dims(lat::Lattice{E,L}) -> (E, L)

Return a tuple (E, L) of the embedding and lattice dimensions of lat

Quantica.dosKPM — Method

dosKPM(h::AbstractMatrix; resolution = 2, kw...)

Compute, using the Kernel Polynomial Method (KPM), the local density of states ρ(ϵ) = ⟨ket|δ(ϵ-h)|ket⟩/⟨ket|ket⟩ for a given ket::AbstractVector and hamiltonian h, or the global density of states ρ(ϵ) = Tr[δ(ϵ-h)] if ket is missing.

If ket is an AbstractMatrix it evaluates the trace over the set of kets in ket (see momentaKPM and its options kw for further details). The result is a tuple of energy points xk::Vector and real ρ::Vector values (any imaginary part in ρ is dropped), where the number of energy points xk is order * resolution, rounded to the closest integer.

dosKPM(μ::MomentaKPM; resolution = 2)

Same as above with momenta μ as input.

dosKPM(h::Hamiltonian; kw...)

Equivalent to dosKPM(bloch(h); kw...) for finite hamiltonians (zero dimensional).

Quantica.energies — Method

energies(s::Spectrum)

Return the energies of s as a Vector

See also

`spectrum`, `states`

Quantica.flatten — Method

flatten(h::Hamiltonian)

Flatten a multiorbital Hamiltonian h into one with a single orbital per site. The associated lattice is flattened also, so that there is one site per orbital for each initial site (all at the same position). Note that in the case of sparse Hamiltonians, zeros in hopping/onsite matrices are preserved as structural zeros upon flattening.

h |> flatten()

Functional form equivalent to flatten(h) of h |> flatten (included for consistency with the rest of the API).

Examples

julia> h = LatticePresets.honeycomb() |> hamiltonian(hopping(@SMatrix[1 2], range = 1/√3, sublats = (:A,:B)), orbitals = (Val(1), Val(2)))
Hamiltonian{<:Lattice} : 2D Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 5 (SparseMatrixCSC, sparse)
  Harmonic size    : 2 × 2
  Orbitals         : ((:a,), (:a, :a))
  Element type     : 2 × 2 blocks (Complex{Float64})
  Onsites          : 0
  Hoppings         : 6
  Coordination     : 3.0

julia> flatten(h)
Hamiltonian{<:Lattice} : 2D Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 5 (SparseMatrixCSC, sparse)
  Harmonic size    : 3 × 3
  Orbitals         : ((:flat,), (:flat,))
  Element type     : scalar (Complex{Float64})
  Onsites          : 0
  Hoppings         : 12
  Coordination     : 4.0

Quantica.hamiltonian — Method

hamiltonian(lat[, model]; orbitals, type)

Create a Hamiltonian by additively applying model::TighbindingModel to the lattice lat (see hopping and onsite for details on building tightbinding models).

The number of orbitals on each sublattice can be specified by the keyword orbitals (otherwise all sublattices have one orbital by default). The following, and obvious combinations, are possible formats for the orbitals keyword:

orbitals = :a                # all sublattices have 1 orbital named :a
orbitals = (:a,)             # same as above
orbitals = (:a, :b, 3)       # all sublattices have 3 orbitals named :a and :b and :3
orbitals = ((:a, :b), (:c,)) # first sublattice has 2 orbitals, second has one
orbitals = ((:a, :b), :c)    # same as above
orbitals = (Val(2), Val(1))  # same as above, with automatic names
orbitals = (:A => (:a, :b), :D => :c) # sublattice :A has two orbitals, :D and rest have one
orbitals = :D => Val(4)      # sublattice :D has four orbitals, rest have one

The matrix sizes of tightbinding model must match the orbitals specified. Internally, we define a block size N = max(num_orbitals). If N = 1 (all sublattices with one orbital) the the Hamiltonian element type is type. Otherwise it is SMatrix{N,N,type} blocks, padded with the necessary zeros as required. Keyword type is Complex{T} by default, where T is the number type of lat.

h(ϕ₁, ϕ₂, ...)
h((ϕ₁, ϕ₂, ...))

Build the Bloch Hamiltonian matrix bloch(h, (ϕ₁, ϕ₂, ...)) of a h::Hamiltonian on an LD lattice. (See also bloch! for a non-allocating version of bloch.)

lat |> hamiltonian(model[, funcmodel]; kw...)

Functional hamiltonian form equivalent to hamiltonian(lat, model[, funcmodel]; kw...).

Indexing

Indexing into a Hamiltonian h works as follows. Access the HamiltonianHarmonic matrix at a given dn::NTuple{L,Int} with h[dn]. Assign v into element (i,j) of said matrix with h[dn][i,j] = v or h[dn, i, j] = v. Broadcasting with vectors of indices is and js is supported, h[dn][is, js] = v_matrix.

To add an empty harmonic with a given dn::NTuple{L,Int}, do push!(h, dn). To delete it, do deleteat!(h, dn).

Examples

julia> h = hamiltonian(LatticePresets.honeycomb(), hopping(@SMatrix[1 2; 3 4], range = 1/√3), orbitals = Val(2))
Hamiltonian{<:Lattice} : 2D Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 5 (SparseMatrixCSC, sparse)
  Harmonic size    : 2 × 2
  Orbitals         : ((:a, :a), (:a, :a))
  Element type     : 2 × 2 blocks (Complex{Float64})
  Onsites          : 0
  Hoppings         : 6
  Coordination     : 3.0

julia> push!(h, (3,3)) # Adding a new Hamiltonian harmonic (if not already present)
Hamiltonian{<:Lattice} : 2D Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 6 (SparseMatrixCSC, sparse)
  Harmonic size    : 2 × 2
  Orbitals         : ((:a, :a), (:a, :a))
  Element type     : 2 × 2 blocks (Complex{Float64})
  Onsites          : 0
  Hoppings         : 6
  Coordination     : 3.0

julia> h[3,3][1,1] = @SMatrix[1 2; 2 1]; h[3,3] # element assignment
2×2 SparseArrays.SparseMatrixCSC{StaticArrays.SArray{Tuple{2,2},Complex{Float64},2,4},Int64} with 1 stored entry:
  [1, 1]  =  [1.0+0.0im 2.0+0.0im; 2.0+0.0im 1.0+0.0im]

julia> h[3,3][[1,2],[1,2]] .= rand(SMatrix{2,2,Float64}, 2, 2) # Broadcast assignment
2×2 view(::SparseArrays.SparseMatrixCSC{StaticArrays.SArray{Tuple{2,2},Complex{Float64},2,4},Int64}, [1, 2], [1, 2]) with eltype StaticArrays.SArray{Tuple{2,2},Complex{Float64},2,4}:
 [0.271152+0.0im 0.921417+0.0im; 0.138212+0.0im 0.525911+0.0im]  [0.444284+0.0im 0.280035+0.0im; 0.565106+0.0im 0.121869+0.0im]
 [0.201126+0.0im 0.912446+0.0im; 0.372099+0.0im 0.931358+0.0im]  [0.883422+0.0im 0.874016+0.0im; 0.296095+0.0im 0.995861+0.0im]

julia> hopfunc(;k = 0) = hopping(k); hamiltonian(LatticePresets.square(), onsite(1) + hopping(2), hopfunc) # Parametric Hamiltonian
Parametric Hamiltonian{<:Lattice} : 2D Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 5 (SparseMatrixCSC, sparse)
  Harmonic size    : 1 × 1
  Orbitals         : ((:a,),)
  Elements         : scalars (Complex{Float64})
  Onsites          : 1
  Hoppings         : 4
  Coordination     : 4.0

Quantica.hopping! — Method

hopping!(f; forcehermitian = true, kw...)
hopping!(f, hoppingselector(; kw...); forcehermitian = true)

Create an ElementModifier, to be used with parametric, that applies f to hoppings specified by hoppingselector(; kw...). The form of f may be f = (t; kw...) -> ... or f = (t, r, dr; kw...) -> ... if the modification is position (r, dr) dependent. The former is naturally more efficient, as there is no need to compute the positions of the two sites involved in each hopping. If forcehermitian == true the modification is forced to yield a hermitian Hamiltonian if the original was hermitian.

See also:

`onsite!`, `parametric`

Quantica.hopping — Method

hopping(t; forcehermitian = true, range = 1, kw...)
hopping(t, hoppingselector(; range = 1, kw...); forcehermitian = true)

Create an TightbindingModelTerm that applies a hopping t to a Lattice when creating a Hamiltonian with hamiltonian. A subset of hoppings can be specified with the kw..., see hoppingselector for details. Note that a default range = 1 is assumed.

The hopping amplitude t can be a number, a matrix (preferably SMatrix) or a function of the form (r, dr) -> ... for a position-dependent hopping (r is the bond center, and dr the bond vector). If sublats is specified as a sublattice name pair, or tuple thereof, hopping is only applied between sublattices with said names. If forcehermitian is true, the model will produce an hermitian Hamiltonian.

The dimension of t::AbstractMatrix must match the orbital dimension of applicable sublattices (see also orbitals option for hamiltonian). If t::Number it will be treated as t * I (proportional to identity matrix) when applied to multiorbital sublattices.

TightbindingModelTerms created with onsite or hopping can be added or substracted together to build more complicated TightbindingModels.

Examples

julia> onsite(1) - hopping(2, dn = ((1,2), (0,0)), sublats = :A=>:B)
TightbindingModel{2}: model with 2 terms
  OnsiteTerm{Int64}:
    Sublattices      : any
    Force hermitian  : true
    Coefficient      : 1
  HoppingTerm{Int64}:
    Sublattice pairs : (:A => :B,)
    dn cell distance : ([1, 2], [0, 0])
    Hopping range    : 1.0
    Force hermitian  : true
    Coefficient      : -1

julia> LatticePresets.honeycomb() |> hamiltonian(hopping((r,dr) -> cos(r[1]), sublats = ((:A,:A), (:B,:B))))
Hamiltonian{<:Lattice} : Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 7 (SparseMatrixCSC, sparse)
  Harmonic size    : 2 × 2
  Orbitals         : ((:a,), (:a,))
  Element type     : scalar (Complex{Float64})
  Onsites          : 0
  Hoppings         : 12
  Coordination     : 6.0

See also:

`onsite`, `onsiteselector`, `hoppingselector`

Quantica.hoppingselector — Method

hoppingselector(; region = missing, sublats = missing, dn = missing, range = missing)

Specifies a subset of hoppings in a given hamiltonian. Only hoppings between two sites at positions r₁ = r - dr/2 and r₂ = r + dr, belonging to unit cells at integer distance dn´ and to sublattices s₁ and s₂ will be selected if: region(r, dr) && s in sublats && dn´ in dn && norm(dr) <= range. If any of these is missing it will not be used to constraint the selection.

See also:

`onsiteselector`, `onsite`, `hopping`

Quantica.lattice — Method

lattice([bravais::Bravais,] sublats::Sublat...; dim::Val{E}, type::T, names)

Create a Lattice{E,L,T} with Bravais matrix bravais and sublattices sublats converted to a common E-dimensional embedding space and type T. To override the embedding dimension E, use keyword dim = Val(E). Similarly, override type T with type = T.

The keywords names can be used to rename sublats. Given names can be replaced to ensure that all sublattice names are unique.

See also LatticePresets for built-in lattices.

Examples

julia> lattice(bravais((1, 0)), sublat((0, 0)), sublat((0, Float32(1))); dim = Val(3))
Lattice{3,1,Float32} : 1D lattice in 3D space
  Bravais vectors : ((1.0, 0.0, 0.0),)
  Sublattices     : 2
    Names         : (:A, :B)
    Sites         : (1, 1) --> 2 total per unit cell

julia> LatticePresets.honeycomb(semibounded = 1, names = (:C, :D))
Lattice{2,2,Float64} : 2D lattice in 2D space
  Bravais vectors : ((0.5, 0.866025), (-0.5, 0.866025))
    Semibounded   : (1)
  Sublattices     : 2
    Names         : (:C, :D)
    Sites         : (1, 1) --> 2 total per unit cell

See also:

LatticePresets, bravais, sublat, supercell, intracell

Quantica.marchingmesh — Method

marchingmesh(npoints::Integer...; axes = 1.0 * I, shift = missing)

Creates a L-dimensional marching-tetrahedra Mesh. The mesh is confined to the box defined by the rows of axes, shifted by shift (an NTuple or Number) if not missing, and contains npoints[i] along each axis i.

marchingmesh(ranges::AbstractRange...; axes = 1.0 * I, shift = missing)

The same as above, but allows to specify the points in axis i by a range ranges[i], such as e.g. -1.0:0.1:1.0.

Note that the size of axes should match the number L of elements in npoints or ranges. The eltype of points is given by that of ranges or axes.

marchingmesh(h::Hamiltonian{<:Lattice}; npoints = 13, shift = missing)

Equivalent to marchingmesh(ntuple(_ -> range(-π, π; length = npoints), Val(L))...) where L is the dimension of the Hamiltonian's lattice.

External links

Marching tetrahedra (https://en.wikipedia.org/wiki/Marching_tetrahedra) in Wikipedia

Quantica.momentaKPM — Function

momentaKPM(h::AbstractMatrix, A = I; ket = missing, order = 10, randomkets = 1, bandrange = missing)

Compute the Kernel Polynomial Method (KPM) momenta μ_n = ⟨ket|T_n(h) A|ket⟩/⟨ket|ket⟩ where T_n(x) is the Chebyshev polynomial of order n, for a given ket::AbstractVector, hamiltonian h, and observable A. If ket is missing, momenta are computed by means of a stochastic trace μ_n = Tr[A T_n(h)] ≈ ∑ₐ⟨a|A T_n(h)|a⟩/N over N = randomkets normalized random |a⟩. Furthermore, the trace over a specific set of kets can also be computed; in this case ket::AbstractMatrix must be a matrix where the columns are the kets involved in the calculation.

The order of the Chebyshev expansion is order. The bandbrange = (ϵmin, ϵmax) should completely encompass the full bandwidth of hamiltonian. If missing it is computed automatically using ArnoldiMethods (must be loaded).

Example

julia> h = LatticePresets.cubic() |> hamiltonian(hopping(1)) |> unitcell(region = RegionPresets.sphere(10));

julia> momentaKPM(bloch(h), bandrange = (-6,6))
Quantica.MomentaKPM{Float64}([0.9594929736144973, -0.005881595972403821, -0.4933354572913581, 0.00359537502632597, 0.09759451291347333, -0.0008081453185250322, -0.00896262538765363, 0.00048205637037715177, -0.0003705198310034668, 9.64901673962623e-20, 9.110915988898614e-18], (0.0, 6.030150753768845))

Quantica.onsite! — Method

onsite!(f; forcehermitian = true, kw...)
onsite!(f, onsiteselector(; kw...); forcehermitian = true)

Create an ElementModifier, to be used with parametric, that applies f to onsite energies specified by onsiteselector(; kw...). The form of f may be f = (o; kw...) -> ... or f = (o, r; kw...) -> ... if the modification is position (r) dependent. The former is naturally more efficient, as there is no need to compute the positions of each onsite energy. If forcehermitian == true the modification is forced to yield a hermitian Hamiltonian if the original was hermitian.

See also:

`hopping!`, `parametric`

Quantica.onsite — Method

onsite(o; forcehermitian = true, kw...)
onsite(o, onsiteselector(; kw...); forcehermitian = true)

Create an TightbindingModelTerm that applies an onsite energy o to a Lattice when creating a Hamiltonian with hamiltonian. A subset of sites can be specified with the kw..., see onsiteselector for details.

The onsite energy o can be a number, a matrix (preferably SMatrix) or a function of the form r -> ... for a position-dependent onsite energy. If forcehermitian is true, the model will produce an hermitian Hamiltonian.

The dimension of o::AbstractMatrix must match the orbital dimension of applicable sublattices (see also orbitals option for hamiltonian). If o::Number it will be treated as o * I (proportional to identity matrix) when applied to multiorbital sublattices.

TightbindingModelTerms created with onsite or hopping can be added or substracted together to build more complicated TightbindingModels.

Examples

julia> onsite(1, sublats = (:A,:B)) - hopping(2, sublats = :A=>:A)
TightbindingModel{2}: model with 2 terms
  OnsiteTerm{Int64}:
    Sublattices      : (:A, :B)
    Force hermitian  : true
    Coefficient      : 1
  HoppingTerm{Int64}:
    Sublattice pairs : (:A => :A,)
    dn cell distance : any
    Hopping range    : 1.0
    Force hermitian  : true
    Coefficient      : -1

julia> LatticePresets.honeycomb() |> hamiltonian(onsite(r->@SMatrix[1 2; 3 4]), orbitals = Val(2))
Hamiltonian{<:Lattice} : Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 1 (SparseMatrixCSC, sparse)
  Harmonic size    : 2 × 2
  Orbitals         : ((:a, :a), (:a, :a))
  Element type     : 2 × 2 blocks (Complex{Float64})
  Onsites          : 2
  Hoppings         : 0
  Coordination     : 0.0

See also:

`hopping`, `onsiteselector`, `hoppingselector`

Quantica.onsiteselector — Method

onsiteselector(; region = missing, sublats = missing)

Specifies a subset of onsites energies in a given hamiltonian. Only sites at position r in sublattice with name s::NameType will be selected if region(r) && s in sublats is true. Any missing region or sublat will not be used to constraint the selection.

See also:

`hoppingselector`, `onsite`, `hopping`

Quantica.optimize! — Method

optimize!(h::Hamiltonian)

Prepare a sparse Hamiltonian h to increase the performance of subsequent calls to bloch(h, ϕs...) and bloch!(matrix, h, ϕs...) by minimizing memory reshufflings. It also adds missing structural zeros to the diagonal to enable shifts by α*I (for shift-and-invert methods).

No optimization will be performed on non-sparse Hamiltonians, or those defined on Superlattices, for which Bloch Hamiltonians are lazily evaluated.

Note that when calling similarmatrix(h) on a sparse h, optimize! is called first.

See also:

bloch, bloch!

Quantica.parametric — Method

parametric(h::Hamiltonian, modifiers::ElementModifier...)

Builds a ParametricHamiltonian that can be used to efficiently apply modifiers to h. modifiers can be any number of onsite!(f;...) and hopping!(f; ...) transformations, each with a set of parameters given as keyword arguments of functions f. The resulting ph::ParamtricHamiltonian can be used to produced the modified Hamiltonian simply by calling it with those same parameters as keyword arguments.

Note 1: for sparse h, parametric only modifies existing onsites and hoppings in h, so be sure to add zero onsites and/or hoppings to h if they are originally not present but you need to apply modifiers to them.

Note 2: optimize!(h) is called prior to building the parametric Hamiltonian. This can lead to extra zero onsites and hoppings being stored in sparse hs.

h |> parametric(modifiers::ElementModifier...)

Function form of parametric, equivalent to parametric(h, modifiers...).

Examples

julia> ph = LatticePresets.honeycomb() |> hamiltonian(onsite(0) + hopping(1, range = 1/√3)) |> unitcell(10) |> parametric(onsite!((o; μ) -> o - μ))
ParametricHamiltonian{<:Lattice} : Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 5 (SparseMatrixCSC, sparse)
  Harmonic size    : 200 × 200
  Orbitals         : ((:a,), (:a,))
  Element type     : scalar (Complex{Float64})
  Onsites          : 200
  Hoppings         : 600
  Coordination     : 3.0
  Param Modifiers  : 1

julia> ph(; μ = 2)
Hamiltonian{<:Lattice} : Hamiltonian on a 2D Lattice in 2D space
  Bloch harmonics  : 5 (SparseMatrixCSC, sparse)
  Harmonic size    : 200 × 200
  Orbitals         : ((:a,), (:a,))
  Element type     : scalar (Complex{Float64})
  Onsites          : 200
  Hoppings         : 600
  Coordination     : 3.0

# See also
    `onsite!`, `hopping!`

Quantica.similarmatrix — Method

similarmatrix(h::Hamiltonian; optimize = true)

Create an uninitialized matrix of the same type of the Hamiltonian's matrix, calling optimize!(h) first if optimize = true to produce an optimal work matrix in the sparse case.

Quantica.sites — Method

sites(lat[, sublat::Int])

Extract the positions of all sites in a lattice, or in a specific sublattice

Quantica.spectrum — Method

spectrum(h; method = defaultmethod(h))

Compute the spectrum of a 0D Hamiltonian h (or alternatively of the bounded unit cell of a finite dimensional h) using one of the following methods

method                    diagonalization function
--------------------------------------------------------------
LinearAlgebraPackage()     LinearAlgebra.eigen!
ArpackPackage()            Arpack.eigs (must be `using Arpack`)

The energies and eigenstates in the resulting s::Spectrum object can be accessed with energies(s) and states(s)

See also

`energies`, `states`, `bandstructure`

Quantica.states — Method

states(bs::Bandstructure, i)

Return the states of each vertex of the i-th band in bs, in the form of a Matrix of size (nψ, nk), where nψ is the length of each state vector, and nk the number of vertices.

Quantica.states — Method

states(s::Spectrum)

Return the states of s as the columns of a Matrix

See also

`spectrum`, `energies`

Quantica.sublat — Method

sublat(sites...; name::Symbol, orbitals = (:noname,))
sublat(sites::Vector{<:SVector}; name::Symbol, orbitals = (:noname,))

Create a Sublat{E,T,D} that adds a sublattice, of name name, with sites at positions sites in E dimensional space, each of which hosts D different orbitals, with orbital names specified by orbitals. Sites can be entered as tuples or SVectors.

Examples

julia> sublat((0.0, 0), (1, 1), (1, -1), name = :A, orbitals = (:upspin, :downspin))
Sublat{2,Float64,2} : sublattice in 2D space with 2 orbitals per site
  Sites    : 3
  Name     : :A
  Orbitals : (:upspin, :downspin)

Quantica.supercell — Method

supercell(lat::AbstractLattice{E,L}, v::NTuple{L,Integer}...; region = missing)
supercell(lat::AbstractLattice{E,L}, sc::SMatrix{L,L´,Int}; region = missing)

Generates a Superlattice from an L-dimensional lattice lat with Bravais vectors br´= br * sc, where sc::SMatrix{L,L´,Int} is the integer supercell matrix with the L´ vectors vs as columns. If no v are given, the superlattice will be bounded. If lat has semibounded axes, these cannot be mixed with any other axes (they can only be removed, kept intact or scaled by a factor, see below).

Only sites at position r such that region(r) == true will be included in the supercell. If region is missing, a Bravais unit cell perpendicular to the v axes will be selected for the L-L´ non-periodic directions.

supercell(lattice::AbstractLattice{E,L}, factor::Integer; region = missing)

Calls supercell with a uniformly scaled sc = SMatrix{L,L}(factor * I)

supercell(lattice::AbstractLattice, factors::Integer...; region = missing)

Calls supercell with different scaling along each Bravais vector (diagonal supercell with factors along the diagonal)

lat |> supercell(v...; kw...)

Functional syntax, equivalent to `supercell(lat, v...; kw...)

supercell(h::Hamiltonian, v...; kw...)

Promotes the Lattice of h to a Superlattice without changing the Hamiltonian itself, which always refers to the unitcell of the lattice.

Examples

julia> supercell(LatticePresets.honeycomb(), region = RegionPresets.circle(300))
Superlattice{2,2,Float64,0} : 2D lattice in 2D space, filling a 0D supercell
  Bravais vectors : ((0.5, 0.866025), (-0.5, 0.866025))
  Sublattices     : 2
    Names         : (:A, :B)
    Orbitals      : ((:noname,), (:noname,))
    Sites         : (1, 1) --> 2 total per unit cell
  Supercell{2,0} for 0D superlattice of the base 2D lattice
    Supervectors  : ()
    Supersites    : 652966

julia> supercell(LatticePresets.triangular(), (1,1), (1, -1))
Superlattice{2,2,Float64,2} : 2D lattice in 2D space, filling a 2D supercell
  Bravais vectors : ((0.5, 0.866025), (-0.5, 0.866025))
  Sublattices     : 1
    Names         : (:A)
    Orbitals      : ((:noname,),)
    Sites         : (1) --> 1 total per unit cell
  Supercell{2,2} for 2D superlattice of the base 2D lattice
    Supervectors  : ((1, 1), (1, -1))
    Supersites    : 2

julia> LatticePresets.square() |> supercell(3)
Superlattice{2,2,Float64,2} : 2D lattice in 2D space, filling a 2D supercell
  Bravais vectors : ((1.0, 0.0), (0.0, 1.0))
  Sublattices     : 1
    Names         : (:A)
    Orbitals      : ((:noname,),)
    Sites         : (1) --> 1 total per unit cell
  Supercell{2,2} for 2D superlattice of the base 2D lattice
    Supervectors  : ((3, 0), (0, 3))
    Supersites    : 9

See also:

unitcell

Quantica.transform! — Method

transform!(h::Hamiltonian, f::Function)

Transform the site positions of the Hamiltonian's lattice in place without modifying the Hamiltonian harmonics.

Quantica.transform! — Method

transform!(lat::Lattice, f::Function)

Transform the site positions of lat by applying f to them in place.

Quantica.unitcell — Method

unitcell(lat::Lattice{E,L}, v::NTuple{L,Integer}...; region = missing)
unitcell(lat::Lattice{E,L}, uc::SMatrix{L,L´,Int}; region = missing)

Generates a Lattice from an L-dimensional lattice lat and a larger unit cell, such that its Bravais vectors are br´= br * uc. Here uc::SMatrix{L,L´,Int} is the integer unitcell matrix, with the L´ vectors vs as columns. If no v are given, the new lattice will be bounded.

Only sites at position r such that region(r) == true will be included in the new unitcell. If region is missing, a Bravais unitcell perpendicular to the v axes will be selected for the L-L´ non-periodic directions.

unitcell(lattice::Lattice{E,L}, factor::Integer; region = missing)

Calls unitcell with a uniformly scaled uc = SMatrix{L,L}(factor * I)

unitcell(lattice::Lattice{E,L}, factors::Integer...; region = missing)

Calls unitcell with different scaling along each Bravais vector (diagonal supercell with factors along the diagonal)

unitcell(slat::Superlattice)

Convert Superlattice slat into a lattice with its unit cell matching slat's supercell.

unitcell(h::Hamiltonian, v...; modifiers = (), kw...)

Transforms the Lattice of h to have a larger unitcell, while expanding the Hamiltonian accordingly. The modifiers (a tuple of ElementModifiers, either onsite! or hopping!) will be applied to onsite and hoppings as the hamiltonian is expanded. See onsite! and hopping! for details

Note: for performance reasons, in sparse hamiltonians only the stored onsites and hoppings will be transformed by ElementModifiers, so you might want to add zero onsites or hoppings when building h to have a modifier applied to them later. Note also that additional onsites and hoppings may be stored when calling optimize! or bloch/bloch! on h for the first time.

lat_or_h |> unitcell(v...; kw...)

Functional syntax, equivalent to `unitcell(latorh, v...; kw...)

Examples

julia> unitcell(LatticePresets.honeycomb(), region = RegionPresets.circle(300))
Lattice{2,0,Float64} : 0D lattice in 2D space
  Bravais vectors : ()
  Sublattices     : 2
    Names         : (:A, :B)
    Orbitals      : ((:noname,), (:noname,))
    Sites         : (326483, 326483) --> 652966 total per unit cell

julia> unitcell(LatticePresets.triangular(), (1,1), (1, -1))
Lattice{2,2,Float64} : 2D lattice in 2D space
  Bravais vectors : ((0.0, 1.732051), (1.0, 0.0))
  Sublattices     : 1
    Names         : (:A)
    Orbitals      : ((:noname,),)
    Sites         : (2) --> 2 total per unit cell

julia> LatticePresets.square() |> unitcell(3)
Lattice{2,2,Float64} : 2D lattice in 2D space
  Bravais vectors : ((3.0, 0.0), (0.0, 3.0))
  Sublattices     : 1
    Names         : (:A)
    Orbitals      : ((:noname,),)
    Sites         : (9) --> 9 total per unit cell

julia> supercell(LatticePresets.square(), 3) |> unitcell
Lattice{2,2,Float64} : 2D lattice in 2D space
  Bravais vectors : ((3.0, 0.0), (0.0, 3.0))
  Sublattices     : 1
    Names         : (:A)
    Orbitals      : ((:noname,),)
    Sites         : (9) --> 9 total per unit cell

See also:

supercell

Quantica.vertices — Method

vertices(bs::Bandstructure, i)

Return the vertices (k..., ϵ) of the i-th band in bs, in the form of a Vector{SVector{L+1}}, where L is the lattice dimension.

Quantica.wrap — Method

wrap(h::Hamiltonian, axis::Int; factor = 1)

Build a new Hamiltonian wherein the Bravais axis is wrapped into a loop. If a factor is given, the wrapped hoppings will be multiplied by said factor. This is useful to represent a flux Φ through the loop, if factor = exp(im * 2π * Φ/Φ₀).

h |> wrap(axis; kw...)

Functional form equivalent to wrap(h, axis; kw...).

Examples

julia> LatticePresets.honeycomb() |> hamiltonian(hopping(1, range = 1/√3)) |> unitcell((1,-1), (10, 10)) |> wrap(2)
Hamiltonian{<:Lattice} : 1D Hamiltonian on a 1D Lattice in 2D space
  Bloch harmonics  : 3 (SparseMatrixCSC, sparse)
  Harmonic size    : 40 × 40
  Orbitals         : ((:a,), (:a,))
  Element type     : scalar (Complex{Float64})
  Onsites          : 0
  Hoppings         : 120
  Coordination     : 3.0

Quantica.BoxIterator — Type

BoxIterator(seed::SVector{N,Int}; maxiterations = missing)

Cartesian iterator iter over SVector{N,Int}s (cells) that starts at seed and grows outwards in the form of a box of increasing sides (not necesarily equal) until it encompasses a certain N-dimensional region. To signal that a cell is in the desired region the user calls acceptcell!(iter, cell).

Base.factorial — Method

factorial(n, k)

Compute $n!/k!$.

Quantica.offdiagonal — Method

offdiagonal(model, lat, nsublats::NTuple{N,Int})

Build a restricted version of model that applies only to off-diagonal blocks formed by sublattice groups of size nsublats.

Quantica.permutations — Method

permutations(a, t)

Generate all size t permutations of an indexable object a.

Quantica.permutations — Method

permutations(a)

Generate all permutations of an indexable object a in lexicographic order. Because the number of permutations can be very large, this function returns an iterator object. Use collect(permutations(a)) to get an array of all permutations.

Funding

This work has been partly funded by the Spanish Ministry of Economy and Competitiveness under Grant Nos. FIS2015-65706-P, PCI2018-093026, and the CSIC Intramural Project 201760I086.